// Frontend/ordermanagement/src/graphql/schema.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Decimal: any;
};

export type Address = {
  __typename?: 'Address';
  addressLine1?: Maybe<Scalars['String']>;
  addressLine2?: Maybe<Scalars['String']>;
  city?: Maybe<Scalars['String']>;
  country?: Maybe<Scalars['String']>;
  customerId: Scalars['Int'];
  id: Scalars['Int'];
  state?: Maybe<Scalars['String']>;
};
// ...
export type Customer = {
  __typename?: 'Customer';
  address?: Maybe<Address>;
  contactNumber?: Maybe<Scalars['String']>;
  email?: Maybe<Scalars['String']>;
  firstName?: Maybe<Scalars['String']>;
  id: Scalars['Int'];
  isDeleted: Scalars['Boolean'];
  lastName?: Maybe<Scalars['String']>;
  orders?: Maybe<Array<Maybe<Order>>>;
};
// ...
export type Order = {
  __typename?: 'Order';
  customer?: Maybe<Customer>;
  customerId: Scalars['Int'];
  depositAmount: Scalars['Decimal'];
  description?: Maybe<Scalars['String']>;
  id: Scalars['Int'];
  isDeleted: Scalars['Boolean'];
  isDelivery: Scalars['Boolean'];
  orderDate: Scalars['DateTime'];
  otherNotes?: Maybe<Scalars['String']>;
  status: Status;
  totalAmount: Scalars['Decimal'];
};
// ...
export enum Status {
  Completed = 'COMPLETED',
  Draft = 'DRAFT',
  Pending = 'PENDING',
  Shipped = 'SHIPPED'
}